+++ tests/slmp_minimal_tests.cpp
@@ -959,14 +959,18 @@ void testLinkDirectWriteDeviceGuards() {
 
 void testTargetAndMonitoringTimerHeaders() {
     assert(slmp::module_io::ControlCpu == 0x03D0U);
-    assert(slmp::module_io::ActiveCpu == slmp::module_io::ControlCpu);
-    assert(slmp::module_io::StandbyCpu == 0x03D1U);
-    assert(slmp::module_io::TypeACpu == 0x03D2U);
-    assert(slmp::module_io::TypeBCpu == 0x03D3U);
-    assert(slmp::module_io::Cpu1 == 0x03E0U);
-    assert(slmp::module_io::Cpu2 == 0x03E1U);
-    assert(slmp::module_io::Cpu3 == 0x03E2U);
-    assert(slmp::module_io::Cpu4 == 0x03E3U);
+    assert(slmp::module_io::ControlSystemCpu == slmp::module_io::ControlCpu);
+    assert(slmp::module_io::StandbySystemCpu == 0x03D1U);
+    assert(slmp::module_io::SystemACpu == 0x03D2U);
+    assert(slmp::module_io::SystemBCpu == 0x03D3U);
+    assert(slmp::module_io::MultipleCpu1 == 0x03E0U);
+    assert(slmp::module_io::MultipleCpu2 == 0x03E1U);
+    assert(slmp::module_io::MultipleCpu3 == 0x03E2U);
+    assert(slmp::module_io::MultipleCpu4 == 0x03E3U);
+    assert(slmp::module_io::RemoteHead1 == slmp::module_io::MultipleCpu1);
+    assert(slmp::module_io::RemoteHead2 == slmp::module_io::MultipleCpu2);
+    assert(slmp::module_io::ControlSystemRemoteHead == slmp::module_io::ControlCpu);
+    assert(slmp::module_io::StandbySystemRemoteHead == slmp::module_io::StandbySystemCpu);
     assert(slmp::module_io::ConnectedCpu == 0x03FFU);
     assert(slmp::module_io::Default == slmp::module_io::ConnectedCpu);
     assert(slmp::TargetAddress{}.module_io == slmp::module_io::ConnectedCpu);
@@ -980,7 +984,7 @@ void testTargetAndMonitoringTimerHeaders() {
     slmp::TargetAddress target = {};
     target.network = 0x12;
     target.station = 0x34;
-    target.module_io = slmp::module_io::Cpu2;
+    target.module_io = slmp::module_io::MultipleCpu2;
     target.multidrop = 0x89;
     plc.setTarget(target);
     plc.setMonitoringTimer(0x4321U);